# frontend/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Picasso",
        "version": "0.1.0"
    },
    "paths": {
        "/api/auth/signup/email": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Signup using email and username. Both of them should be unique. This will send magic link for login on the registered email address",
                "operationId": "email_signup_api_auth_signup_email_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailSignupIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User account created and magic link login sent to email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailSignupOut"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Username or email already taken",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConflictErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login/email": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Login using email. This will send magic send magic link for login on the registered email address",
                "operationId": "email_login_api_auth_login_email_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailLoginIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Magic link login sent to email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailLoginOut"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login/email/{token}": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Complete magic link login and then set refresh token and sending access token",
                "operationId": "complete_email_login_api_auth_login_email__token__get",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Magic link login sent to email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompleteEmailLoginOut"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Refresh access token",
                "operationId": "refresh_access_token_api_auth_refresh_get",
                "responses": {
                    "200": {
                        "description": "Refresh access token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RefreshAccessTokenOut"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Logout user",
                "operationId": "logout_user_api_auth_logout_get",
                "responses": {
                    "204": {
                        "description": "User logged out"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Get current user",
                "operationId": "get_logged_in_user_profile_api_auth_me_get",
                "responses": {
                    "200": {
                        "description": "Get current user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoggedInUserProfile"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BadRequestErrorResponse": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/PydanticErrorResponse"
                    },
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                ],
                "title": "BadRequestErrorResponse"
            },
...
            "CompleteEmailLoginOut": {
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "title": "Accesstoken"
                    },
                    "refreshToken": {
                        "type": "string",
                        "title": "Refreshtoken"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["accessToken", "refreshToken", "user"],
                "title": "CompleteEmailLoginOut"
            },
            "ConflictErrorResponse": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "title": "Reason"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": ["reason", "message"],
                "title": "ConflictErrorResponse"
            },
            "EmailLoginIn": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["email"],
                "title": "EmailLoginIn"
            },
            "EmailLoginOut": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["message"],
                "title": "EmailLoginOut"
            },
            "EmailSignupIn": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3,
                        "title": "Username"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["email", "username"],
                "title": "EmailSignupIn"
            },
            "EmailSignupOut": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["message"],
                "title": "EmailSignupOut"
            },
...
            "ErrorResponse": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "title": "Reason"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": ["reason", "message"],
                "title": "ErrorResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "InternalServerErrorResponse": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "title": "Reason"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": ["reason", "message"],
                "title": "InternalServerErrorResponse"
            },
            "LoggedInUserProfile": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["user"],
                "title": "LoggedInUserProfile"
            },
            "PydanticErrorResponse": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "title": "Reason"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "errors": {
                        "items": {},
                        "type": "array",
                        "title": "Errors"
                    }
                },
                "type": "object",
                "required": ["reason", "message", "errors"],
                "title": "PydanticErrorResponse"
            },
            "RefreshAccessTokenOut": {
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "title": "Accesstoken"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["accessToken"],
                "title": "RefreshAccessTokenOut"
            },
...
            "UnauthorizedErrorResponse": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "title": "Reason"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": ["reason", "message"],
                "title": "UnauthorizedErrorResponse"
            },
...
            "User": {
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Userid"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3,
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "profilePicURL": {
                        "type": "string",
                        "minLength": 1,
                        "format": "uri",
                        "title": "Profilepicurl"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": ["userId", "username", "email", "profilePicURL"],
                "title": "User"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    }
}